split report classes & comment out NutritionReport
